fix typos/move generate script
